==> code/ntbks/05_pandas/04_advanced_topics_1.ipynb <==
### Working with missing values

It occurs rather frequently that DataFrames are incomplete regarding the content of all data rows. There will probably be no occasion in which complete rows are entered empty, however it may very well occur that some part of an observation is missing.<br>
The question is how to deal with this situation? Let's have a look at the following DataFrame.

import pandas as pd
import numpy as np
import os
from pathlib import Path
data_dir = os.path.join(Path(os.getcwd()).parents[2], "data", "ntbk_data", "04_data")
data = {'col1': [1,2,3,None,5], 'col2' : ['a',np.nan,'c','d','b'], 'col3' : [10,20,30,np.nan,np.nan], 
        'col4' : [np.nan,np.nan,np.nan,np.nan,0.6], 'col5' : [np.nan,np.nan,np.nan,np.nan,np.nan], 
        'col6' : ['a','b','c','d','e'], 'col7' : [10,25,40,np.nan,15]}

df = pd.DataFrame.from_dict(data)

df

df.isnull()

df.col1.isnull()#return a Series

As you can see, we entered some 'missing' values on purpose, in this case, the Python __None__ type as well as the __np.nan__ type are converted and handled as __np.nan__ in the pandas DataFrame. The first thing you should do is check the DataFrame for these values so you have an overview of the completeness of your data.
Of course, you can use the __.describe()__ function here and compare the values but checking per column is more thorough.

#combining isnull() with sum() results in a nice overview
df.isnull().sum()

### Removing  data from a DataFrame


What to do if data is missing in a DataFrame? One way is to remove observations (rows) or attributes (columns) completely. This can be done if you are sure that it does not harm your information harvesting overall!<br>
Have a look at our DataFrame - it is easy to see that column 5 does not carry any valuable informartion whatsoever - let's remove it!

df.drop('col5', axis = 1, inplace=True)
#in such a case you may want to know, where the data came from after all or if there was an import error!

df

One may argue, whether there is much information contained in __col4__ as well, but let's leave it for the time being.

#this command removes any column where a np.nan value is found; 
#omitting the axis parameter results in the same behaviour for rows
df1 = df.dropna(axis = 1)
df1

### Imputing data

We can try to fill in the missing data using the __fillna()__ function. This makes sense for numerical data. Let's try this for a single column.

df.col1.fillna(4.0, inplace=True)
df

We don't always want to guess the correct value, neither do we want to inspect each entry specifically, nor is this feasible if the DataFrame is large! A good measure may be to use the mean value in a column. Whether or not this makes sense depends on your knowledge of the data!

print(df.col7.mean())

df.col7.fillna(df.col7.mean(), inplace=True)
df

df.describe()

Using __describe()__ on the column gives a description on the numerical data in the DataFrame:
- count: the number of non-null elements in this column
- mean: the mean value of the column
- std: the standard deviation
- min/max: min and max value respectively
- 25%/50%/75%: The respective percentile

df.col2.describe()#applied on non-numerical data

Applied on a column, we can retrieve:
- count: the number of values
- the number of unique values
- the top value
- the frequency of the top value

df.col7.describe()

### Correlation in data

Here is a link to some [housing data](https://www.kaggle.com/schirmerchad/bostonhoustingmlnd) in the Boston suburbian area. The column __RM__ describes the size in square metres per real estate, the __MEDV__ column refers to the value of the object in question.

dfh = pd.read_csv(os.path.join(data_dir,'housing.csv'), encoding= 'utf-8')
dfh.head()

Using the __corr()__ function we can quickly establish correlation factors for all variables and it is obvious to see that there is a positive correlation between the size of an object and its value, as we expected!

dfh.corr()

### Working with data


Let's work a bit with a concrete data set and apply our knowledge about data engineering / data wrangling to answer some conrete questions! Here is a [link](https://www.kaggle.com/russellyates88/suicide-rates-overview-1985-to-2016) to a data set describing suicides rates in various countries. We import the data into a DataFrame and take an initial glimpse at it.

dfS = pd.read_csv(os.path.join(data_dir, 'suicide_data.csv'))

#rename some column names
dfS.rename(columns = {' gdp_for_year ($) ':'gdp_year', 'gdp_per_capita ($)':'gdp_cap'}, inplace = True)

#if you are unsure about the meaning of the specified data, look it up under the link!
dfS.head()

dfS.describe()

__Grouping__ allows for aggregation over certain columns, e.g. iw want to know all suicide values per country, we can sum all the entries with respect to each country. 

dfGroup = dfS.groupby(['country']).sum(numeric_only=True)
dfGroup

Other forms of aggregation are:
- count: number of non-null observations
- sum:	Sum of values
- mean:	Mean of values
- mad:	Mean absolute deviation
- median: Arithmetic median of values
- min:	Minimum
- max:	Maximum
- mode:	Mode
- abs:	Absolute Value
- prod:	Product of values
- std:	standard deviation
- var:  variance
- sem: standard error of the mean
- quantile:	Sample quantile (value at %)
- cumsum:	Cumulative sum
- cumprod:	Cumulative product
- cummax:	Cumulative maximum
- cummin:	Cumulative minimum

If we want to retrieve parts from our DataFrame we can achieve this by using comparion: Certain values of columns which match our defined criteria.   

#retrieve a DataFrame listing all rows which match the string 'Germany' in the column country
#or: give me all the data regarding Germany
dfG = dfS[dfS.country == 'Germany']
dfG.head()

dfG.describe()

#TASK: retrieve all data of male persons in Germany from the year 2000
dfS.loc[(dfS.country == 'Germany') & (dfS.sex == 'male') & (dfS.year == 2000)]


#TASK: retrieve all data of male persons in Germany from the year 2000
t1 = dfG[(dfG.sex == 'male') & (dfG.year == 2000)]
t1

t1.describe()

Task: Retrieve all data onwards from 2000 regarding [Generation X](https://en.wikipedia.org/wiki/Generation_X). What are the countries involved?

gen_x_older2000 = dfS.loc[(dfS.year > 2000) & (dfS.generation == 'Generation X')]
gen_x_older2000.country.unique()

t2 = dfS[(dfS.year > 2000) & (dfS.generation == 'Generation X')]
t2.head()

t2.country.unique()

#Task: retrieve all data from the DataFrame regarding Generation X in Germany between 2000 and 2010 (excluding both)

dfS.loc[(dfS.generation == 'Generation X') & (dfS.country == 'Germany') & (dfS.year > 2000) & (dfS.year < 2010)]






#Task: retrieve all data from the DataFrame regarding Generation X in Germany between 2000 and 2010 (excluding both)
t3 = t2[(t2.country == 'Germany') & (t2.year > 2000) & (t2.year < 2010)]
t3

#TASK (based on t2): List the suicide rates of the richest (per capita) 10% of all the countries between [2000,2010]
#make use of the .quantile() function

t2_richest_suic_rates = t2.loc[(t2.year > 2000) & (t2.year < 2010) & (t2.gdp_cap > t2.gdp_cap.quantile(0.9)), ['country', 'suicides/100k pop']]

print(t2_richest_suic_rates.head(10))
t2_richest_suic_rates.country.unique()


#TASK (based on t2): List the suicide rates of the richest (per capita) 10% of all the countries between [2000,2010]
#make use of the .quantile() function
top10 = t2[(t2.year > 2000) & (t2.year < 2010) & (t2.gdp_cap > t2.gdp_cap.quantile(0.9))]
top10.country.unique()

top10.head()

#TASK: solve the same question for the poorest 10%


t2_poor_suic_rates = t2.loc[(t2.year > 2000) & (t2.year < 2010) & (t2.gdp_cap < t2.gdp_cap.quantile(0.1)), ['country', 'suicides/100k pop']]

print(t2_poor_suic_rates.head(10))
t2_poor_suic_rates.country.unique()



#TASK: In Germany - for GenX, are the suicides rates higher for men or women?

t2.loc[t2.country == 'Germany'].groupby(['sex']).mean(numeric_only=True).loc[:, 'suicides/100k pop']







gM = t3[t3.sex=='male']
gM.suicides_no.sum()

gF = t3[t3.sex=='female']
gF.suicides_no.sum()

#TASK: Which generation has the highest abolute suicide rates in Germany?

dfG.groupby(['generation', 'year'], as_index = False)['suicides_no'].sum().groupby(['generation']).max()
#dfS.loc[dfS.country == 'Germany'].groupby(['generation']).max(numeric_only=True).loc[:,'suicides/100k pop']








dfG.head()

df1 = dfG.groupby(['generation','year'], as_index=False)['suicides_no'].sum().groupby(['generation']).max()
#(['A', 'B'], as_index=False)['C'].sum()
df1
